requirements.txt
djangorestframework==3.14.0
django-filter==22.1
coreapi==2.3.3
pytest-django==4.5.2